pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gitrag"
version = "0.1.0"
description = "Production-grade local RAG system for chatting with GitHub repositories"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "GitRAG" }]

dependencies = [
    # AST parsing
    "tree-sitter>=0.22",
    "tree-sitter-python>=0.21",
    "tree-sitter-javascript>=0.21",
    "tree-sitter-typescript>=0.21",
    "tree-sitter-cpp>=0.22",
    "tree-sitter-java>=0.21",
    # Embeddings
    "sentence-transformers>=2.2",
    "torch>=2.0",
    # Vector store
    "chromadb>=0.4",
    # BM25
    "rank-bm25>=0.2",
    # Reranking
    # (uses sentence-transformers CrossEncoder)
    # LLM
    "openai>=1.0",
    "httpx>=0.25",
    # API
    "fastapi>=0.104",
    "uvicorn>=0.24",
    # Utilities
    "pyyaml>=6.0",
    "rich>=13.0",
    "click>=8.0",
    "tiktoken>=0.5",
    "networkx>=3.0",
    "numpy>=1.24",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "ruff>=0.1",
]

[project.scripts]
gitrag = "gitrag.cli:main"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
